# swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample subscription service API"
  version: "1.0.0"
  title: "Subscriptions"
host: "localhost:3000"
tags:
- name: "subscription"
  description: "Subscription operations"
schemes:
- "http"
paths:
  /subscription:
    post:
      tags:
      - "subscription"
      summary: "Register a new subscription"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "x-access-token"
        in: "header"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Subscription object to register"
        required: true
        schema:
          $ref: "#/definitions/Subscription"
      responses:
        "200":
          description: "OK"
        "405":
          description: "Invalid input"
        "500":
          description: "Internal server error"
      security:
      - jwt: []
    get:
      tags:
      - "subscription"
      summary: "Get all the active subscriptions"
      produces:
      - "application/json"
      parameters:
      - name: "x-access-token"
        in: "header"
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/ArrayOfSubscriptions"
        "500":
          description: "Internal server error"
  /subscriptions/{subscriptionId}:
    get:
      tags:
      - "subscription"
      summary: "Find subscription by ID"
      produces:
      - "application/json"
      parameters:
      - name: "x-access-token"
        in: "header"
        required: true
        type: "string"
      - name: "subscriptionId"
        in: "path"
        description: "ID of subscription to return"
        required: true
        type: "integer"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/Subscription"
        "404":
          description: "Not found"
        "500":
          description: "Internal server error"
      security:
      - jwt: []
    delete:
      tags:
      - "subscription"
      summary: "Deletes a subscription"
      produces:
      - "application/json"
      parameters:
      - name: "x-access-token"
        in: "header"
        required: true
        type: "string"
      - name: "subscriptionId"
        in: "path"
        description: "Subscription id to delete"
        required: true
        type: "integer"
      responses:
        "200":
          description: "OK"
        "500":
          description: "Internal server error"
      security:
      - jwt: []
securityDefinitions:
  jwt:
    type: "apiKey"
    name: "x-access-token"
    in: "header"
definitions:
  Subscription:
    type: "object"
    required:
      - email
      - date_of_birth
      - consent
      - newsletter_id
    properties:
      email:
        type: "string"
        default: "test@test.com"
      firstname:
        type: "string"
        default: "test"
      gender:
        type: "string"
        default: "test"
      date_of_birth:
        type: "string"
        format: "date-time"
      consent:
        type: "boolean"
      newsletter_id:
        type: "integer"
  ArrayOfSubscriptions:
    type: "array"
    items:
      $ref: "#/definitions/Subscription"